EEPROM storage abstraction
We introduce `driver::storage::AVREEPROM`, and a `storage()` method in
`kaleidoscope::Hardware`. The idea is that each board will pull in the storage
driver they need, and we access it through them. This will allow us to support
boards that only have emulated EEPROM.

All callers that used EEPROM directly were updated to use
`KeyboardHardware.storage()` instead, which provides a very similar interface.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>