Introduce a commit() method to the storage driver
While we do not need it for AVR, we will, in the future. It's easier to
introduce it now, though. The purpose of commit() is to really write the data to
storage. We don't want to do it implicitly on .put() or .write(), because that
can potentially wear out storage faster than comitting in larger blocks.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>